pyproject.toml
[project]
name = "lightrag-fefta"
version = "0.1.0"
description = "LightRag FEFTA Project"
authors = [
    {name = "Kou"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "lightrag-hku (>=1.3.1,<2.0.0)",
    "numpy (>=1.22.4,<1.29.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pipmaster (>=0.5.4,<0.6.0)",
    "future (>=1.0.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "textract (>=1.6.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[[tool.poetry.packages]]
include = "lightrag_fefta"
from = "src"

[tool.poetry.scripts]
build = "lightrag_fefta.build_rag:main"
query = "lightrag_fefta.query_rag:main"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
